requirements.txt
confusable-homoglyphs==3.0.0
Django==2.0.4
django-multiselectfield==0.1.8
django-registration==2.4.1
factory-boy==2.10.0
Faker==0.8.13
Pillow==5.1.0
psycopg2==2.7.4
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
